feat(api): updated duplicate ip report error handling (#571)
BREAKING CHANGE: If you report duplicate IPs within a 15 minutes time frame, the response object
sent by the API will be different. The result contains a detail/source/status definition, with the
code 429, and the headers will also have a statusText as 'Too Many Requests'. There is a sample
definition on the api test file.

fix #570 #568 #567 #566